agents/requirements.txt
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
eth-account>=0.11.0
